refactor(runtime): one owner for the grow-only SourceDeath cursor
The `SourceDeath` health watch is a grow-only Vec, so every consumer must skip
past already-surfaced entries or it re-warns all prior deaths (N → N(N+1)/2
lines). `floating/mod.rs` owned a tested `unseen_deaths(deaths, &mut seen)`
helper; `headless_loop_with_signal` re-inlined the same contract via
`skip(deaths_seen)` + `deaths_seen = deaths.len()`. Two homes for one subtle
off-by-one.

Move `unseen_deaths` into `runtime/mod.rs` beside its sibling
`format_source_death` (which already imports `SourceDeath`), carry its unit
test, and route both consumers through it. Per-consumer presentation
(headless `println!` vs floating `warn!`) stays where it belongs; the cursor
invariant now has one owner + one test a future third consumer inherits.
Deletion test: concentrates. No dep or purity change.
