+++ crates/pixtuoid/src/floating/mod.rs
@@ -31,18 +31,6 @@ use crate::runtime::driver::{build_source_set, reducer_task};
 use crate::runtime::{ConnectedSources, RunConfig};
 use window::{FloatingApp, FloatingEvent};
 
-/// The not-yet-surfaced tail of the grow-only `SourceDeath` watch Vec, advancing
-/// `seen` past it — the pure half of the health-bridge dedup ("tracked by count:
-/// the watch Vec only grows", the same contract headless_loop keeps inline).
-fn unseen_deaths<'a>(
-    deaths: &'a [pixtuoid_core::source::manager::SourceDeath],
-    seen: &mut usize,
-) -> &'a [pixtuoid_core::source::manager::SourceDeath] {
-    let start = (*seen).min(deaths.len());
-    *seen = deaths.len();
-    &deaths[start..]
-}
-
 /// Open the floating window and drive it until the user closes it.
 ///
 /// `winit`'s event loop must own the main thread, so the source pipeline runs on a
@@ -162,7 +150,7 @@ pub fn run(cfg: RunConfig) -> Result<()> {
             let mut deaths_seen = 0usize;
             while health_rx.changed().await.is_ok() {
                 let deaths = health_rx.borrow_and_update().clone();
-                for death in unseen_deaths(&deaths, &mut deaths_seen) {
+                for death in crate::runtime::unseen_deaths(&deaths, &mut deaths_seen) {
                     tracing::warn!("pixtuoid floating: source exited: {death:?}");
                 }
             }
@@ -183,31 +171,4 @@ pub fn run(cfg: RunConfig) -> Result<()> {
         .run_app(&mut app)
         .context("running the floating window event loop")?;
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::unseen_deaths;
-    use pixtuoid_core::source::manager::SourceDeath;
-
-    #[test]
-    fn unseen_deaths_yields_each_death_exactly_once() {
-        let mut seen = 0usize;
-        let one = vec![SourceDeath::new("codex", "boom")];
-        assert_eq!(unseen_deaths(&one, &mut seen).len(), 1);
-        assert_eq!(seen, 1);
-
-        // The grow-only Vec gains a second death: only the NEW one surfaces —
-        // the first must not be re-logged on every later change.
-        let two = vec![
-            SourceDeath::new("codex", "boom"),
-            SourceDeath::new("claude-code", "bind"),
-        ];
-        let fresh = unseen_deaths(&two, &mut seen);
-        assert_eq!(fresh.len(), 1);
-        assert_eq!(fresh[0].source, "claude-code");
-
-        // No growth → nothing to log.
-        assert!(unseen_deaths(&two, &mut seen).is_empty());
-    }
 }
